引入SPI机制，序列化器拓展

+++ example-common/src/main/java/com/code/example/common/service/UserService.java
@@ -21,7 +21,7 @@ public interface UserService {
      *
      * @return
      */
-    default short getNumber() {
-        return 1;
+    default int getNumber() {
+        return  1;
     }
 }
